Fix various explicit cast warnings

+++ extern/imgui/imgui-SFML.cpp
@@ -568,10 +568,10 @@ void ProcessEvent(const sf::Event& event) {
     if (s_currWindowCtx->windowHasFocus) {
         switch (event.type) {
         case sf::Event::Resized:
-            io.DisplaySize = ImVec2(event.size.width, event.size.height);
+            io.DisplaySize = ImVec2(static_cast<float>(event.size.width), static_cast<float>(event.size.height));
             break;
         case sf::Event::MouseMoved:
-            io.AddMousePosEvent(event.mouseMove.x, event.mouseMove.y);
+            io.AddMousePosEvent(static_cast<float>(event.mouseMove.x), static_cast<float>(event.mouseMove.y));
             s_currWindowCtx->mouseMoved = true;
             break;
         case sf::Event::MouseButtonPressed: // fall-through
